Add result check to attempt to load torrent file for conditional `rm`
Now only successfully added *.torrent files will be removed, with acknowledgement, while failed additions will be left alone for the user to check.

+++ load-to-transmission-daemon.sh
@@ -13,5 +13,11 @@ fi
 
 for filename in "$1"*.torrent; do
     transmission-remote --authenv --add "$filename"
-    rm "$filename"
+    if [ $? -eq 0 ]; then
+		echo "Removing $filename"
+		rm "$filename"
+	else
+		echo "Retaining $filename"
+	fi
+
 done
